requirements.txt
streamlit
requests
gtts
pygame
sounddevice
wavio
speechrecognition
python-dotenv
deep-translator
openai==0.28.1
